refactor(#573): extract SettingsSheet.vue from App.vue
Extract the settings bottom sheet and all related UI (legal modals,
confirm dialogs, delete account, theme selection, progression toggles,
data export/import, weight goal) into a dedicated SettingsSheet.vue
component. App.vue drops from 1951 to 599 lines, keeping only app
shell responsibilities: auth gating, tab routing, sync status,
onboarding, and global overlays (XP toast, unlock celebration, PR burst).

SettingsSheet communicates with App.vue via v-model for open/close state
and a sign-out event. All store access and business logic that was
previously in App.vue's settings section now lives in SettingsSheet.
